Updated file structure for extensions
