+++ inc/class-header-footer-elementor.php
@@ -372,6 +372,7 @@ class Header_Footer_Elementor {
 
 		// Load the widgets.
 		require HFE_DIR . 'inc/widgets-manager/class-widgets-loader.php';
+		require HFE_DIR . 'inc/widgets-manager/class-extensions-loader.php';
 	}
 
 	/**
