+++ inc/class-hfe-settings-page.php
@@ -32,7 +32,6 @@ class HFE_Settings_Page {
 		add_filter( 'admin_footer_text', [ $this, 'admin_footer_text' ] );
 		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_admin_scripts' ] );
 		add_filter( 'plugin_action_links_' . HFE_PATH, [ $this, 'settings_link' ] );
-		add_action( 'elementor/element/after_section_end', [ $this, 'register_reading_progress_bar' ], 10, 3 );
 
 	}
 
@@ -853,204 +852,6 @@ class HFE_Settings_Page {
 		return array_merge( $custom, (array) $links );
 	}
 
-	/**
-	 * Registers the Reading Progress Bar setting.
-	 *
-	 * This function adds the Reading Progress Bar setting to the settings page.
-	 *
-	 * @since x.x.x
-	 *
-	 * @param \Elementor\Element_Base $element The element instance.
-	 * @param string                  $section_id The section ID.
-	 * @param array                   $args The arguments.
-	 */
-	public function register_reading_progress_bar( $element, $section_id, $args ) {
-		if ( 'section_page_style' === $section_id || 'container_page_style' === $section_id ) {
-			$element->start_controls_section(
-				'section_reading_progress_bar',
-				[
-					'label' => __( 'Reading Progress Bar', 'header-footer-elementor' ),
-					'tab'   => \Elementor\Controls_Manager::TAB_SETTINGS,
-				]
-			);
-
-			$element->add_control(
-				'enable_reading_progress_bar',
-				[
-					'label'        => __( 'Enable Reading Progress Bar', 'header-footer-elementor' ),
-					'type'         => \Elementor\Controls_Manager::SWITCHER,
-					'label_on'     => __( 'Yes', 'header-footer-elementor' ),
-					'label_off'    => __( 'No', 'header-footer-elementor' ),
-					'return_value' => 'yes',
-					'default'      => 'no',
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_global',
-				[
-					'label'        => __( 'Enable Global Display ', 'header-footer-elementor' ),
-					'description'  => __( 'Enabling this option will affect the entire site.', 'header-footer-elementor' ),
-					'type'         => \Elementor\Controls_Manager::SWITCHER,
-					'default'      => 'no',
-					'label_on'     => __( 'Yes', 'header-footer-elementor' ),
-					'label_off'    => __( 'No', 'header-footer-elementor' ),
-					'return_value' => 'yes',
-					'separator'    => 'before',
-					'condition'    => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-			
-			$element->add_control(
-				'hfe_reading_progress_global_display_condition',
-				[
-					'label'     => __( 'Display On', 'header-footer-elementor' ),
-					'type'      => \Elementor\Controls_Manager::SELECT,
-					'default'   => 'all',
-					'options'   => [
-						'posts' => __( 'All Posts', 'header-footer-elementor' ),
-						'pages' => __( 'All Pages', 'header-footer-elementor' ),
-						'all'   => __( 'All Posts & Pages', 'header-footer-elementor' ),
-					],
-					'condition' => [
-						'enable_reading_progress_bar' => 'yes',
-						'hfe_reading_progress_global' => 'yes',
-					],
-					'separator' => 'before',
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_position',
-				[
-					'label'       => __( 'Position', 'header-footer-elementor' ),
-					'type'        => \Elementor\Controls_Manager::SELECT,
-					'default'     => 'top',
-					'label_block' => false,
-					'options'     => [
-						'top'    => __( 'Top', 'header-footer-elementor' ),
-						'bottom' => __( 'Bottom', 'header-footer-elementor' ),
-						'right'  => __( 'Right', 'header-footer-elementor' ),
-						'left'   => __( 'Left', 'header-footer-elementor' ),
-					],
-					'separator'   => 'before',
-					'condition'   => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_height',
-				[
-					'label'      => __( 'Height', 'header-footer-elementor' ),
-					'type'       => \Elementor\Controls_Manager::SLIDER,
-					'size_units' => [ 'px' ],
-					'range'      => [
-						'px' => [
-							'min'  => 0,
-							'max'  => 100,
-							'step' => 1,
-						],
-					],
-					'default'    => [
-						'unit' => 'px',
-						'size' => 5,
-					],
-					'selectors'  => [
-						'{{WRAPPER}} .hfe-reading-progress-wrap .hfe-reading-progress' => 'height: {{SIZE}}{{UNIT}} !important',
-						'{{WRAPPER}} .hfe-reading-progress-wrap .hfe-reading-progress .hfe-reading-progress-fill' => 'height: {{SIZE}}{{UNIT}} !important',
-					],
-					'separator'  => 'before',
-					'condition'  => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_bg_color',
-				[
-					'label'     => __( 'Background Color', 'header-footer-elementor' ),
-					'type'      => \Elementor\Controls_Manager::COLOR,
-					'default'   => '',
-					'selectors' => [
-						'{{WRAPPER}} .hfe-reading-progress' => 'background-color: {{VALUE}}',
-					],
-					'separator' => 'before',
-					'condition' => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_fill_color',
-				[
-					'label'     => __( 'Fill Color', 'header-footer-elementor' ),
-					'type'      => \Elementor\Controls_Manager::COLOR,
-					'default'   => '#1fd18e',
-					'selectors' => [
-						'{{WRAPPER}} .hfe-reading-progress-wrap .hfe-reading-progress .hfe-reading-progress-fill' => 'background-color: {{VALUE}}',
-					],
-					'separator' => 'before',
-					'condition' => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_animation_speed',
-				[
-					'label'      => __( 'Animation Speed', 'header-footer-elementor' ),
-					'type'       => \Elementor\Controls_Manager::SLIDER,
-					'size_units' => [ 'px' ],
-					'range'      => [
-						'px' => [
-							'min'  => 0,
-							'max'  => 1000,
-							'step' => 1,
-						],
-					],
-					'default'    => [
-						'unit' => 'px',
-						'size' => 50,
-					],
-					'selectors'  => [
-						'{{WRAPPER}} .hfe-reading-progress-wrap .hfe-reading-progress .hfe-reading-progress-fill' => 'transition: width {{SIZE}}ms ease;',
-					],
-					'separator'  => 'before',
-					'condition'  => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->add_control(
-				'hfe_reading_progress_bar_disable_on',
-				[
-					'label'     => __( 'Disable On ', 'header-footer-elementor' ),
-					'type'      => \Elementor\Controls_Manager::SELECT,
-					'multiple'  => true,
-					'options'   => [
-						'desktop' => __( 'Desktop', 'header-footer-elementor' ),
-						'tablet'  => __( 'Tablet', 'header-footer-elementor' ),
-						'mobile'  => __( 'Mobile', 'header-footer-elementor' ),
-					],
-					'default'   => [ 'desktop' ],
-					'separator' => 'before',
-					'condition' => [
-						'enable_reading_progress_bar' => 'yes',
-					],
-				]
-			);
-
-			$element->end_controls_section();
-		}
-	}
 }
 
 new HFE_Settings_Page();
